adds abstract sandbox security manager
