+++ script/src/main/java/org/opentestsystem/rdw/script/ScriptPipelineConfiguration.java
@@ -9,6 +9,7 @@ import org.opentestsystem.rdw.script.impl.ResourceLoaderScriptSource;
 import org.opentestsystem.rdw.script.publishing.ArchiveBackedPublishedPipelineRepository;
 import org.opentestsystem.rdw.script.publishing.PublishedPipelineRepository;
 import org.opentestsystem.rdw.script.security.DefaultSandboxPermissions;
+import org.opentestsystem.rdw.script.security.DefaultSandboxSecurityManager;
 import org.opentestsystem.rdw.script.security.SandboxPermissions;
 import org.opentestsystem.rdw.script.security.SandboxSecurityManager;
 import org.springframework.context.annotation.Bean;
@@ -52,7 +53,7 @@ public class ScriptPipelineConfiguration {
 
     @Bean
     public SandboxSecurityManager sandboxSecurityManager(final SandboxPermissions permissions) {
-        return new SandboxSecurityManager(permissions);
+        return new DefaultSandboxSecurityManager(permissions);
     }
 
     @Bean
