+++ script/src/main/java/org/opentestsystem/rdw/script/impl/DefaultConfigurationCompiler.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.script.impl;
 
 import java.util.HashMap;
 import java.util.Map;
+
 import org.opentestsystem.rdw.script.ConfigurationCompiler;
 
 public class DefaultConfigurationCompiler implements ConfigurationCompiler {
@@ -13,8 +14,8 @@ public class DefaultConfigurationCompiler implements ConfigurationCompiler {
      * @return enhanced properties to be used as configuration for a script
      */
     @Override
-    public Map<String,Object> compile(final Map<String,String> properties) {
-        final Map<String,Object> configuration = new HashMap<>();
+    public Map<String, Object> compile(final Map<String, String> properties) {
+        final Map<String, Object> configuration = new HashMap<>();
         if (properties != null) {
             configuration.putAll(properties);
         }
