+++ script/src/test/java/org/opentestsystem/rdw/script/SandboxTest.java
@@ -8,6 +8,7 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.opentestsystem.rdw.script.security.DefaultSandboxPermissions;
+import org.opentestsystem.rdw.script.security.DefaultSandboxSecurityManager;
 import org.opentestsystem.rdw.script.security.SandboxSecurityManager;
 
 import java.io.FileOutputStream;
@@ -40,7 +41,7 @@ public class SandboxTest {
 
     @BeforeClass
     public static void changeSecurityPolicy() {
-        sandboxSecurityManager = new SandboxSecurityManager(new DefaultSandboxPermissions());
+        sandboxSecurityManager = new DefaultSandboxSecurityManager(new DefaultSandboxPermissions());
         sandboxSecurityManager.disable();
         System.setSecurityManager(sandboxSecurityManager);
     }
