+++ script/src/test/java/org/opentestsystem/rdw/script/XmlPipelineTest.java
@@ -16,6 +16,7 @@ import org.opentestsystem.rdw.script.impl.EmptyPropertyResolver;
 import org.opentestsystem.rdw.script.impl.ResourceLoaderScriptSource;
 import org.opentestsystem.rdw.script.publishing.PublishedPipelineRepository;
 import org.opentestsystem.rdw.script.security.DefaultSandboxPermissions;
+import org.opentestsystem.rdw.script.security.DefaultSandboxSecurityManager;
 import org.opentestsystem.rdw.script.security.SandboxSecurityManager;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
@@ -66,7 +67,7 @@ public class XmlPipelineTest {
 
     @BeforeClass
     public static void setupSecurityManager() {
-        sandboxSecurityManager = new SandboxSecurityManager(new DefaultSandboxPermissions());
+        sandboxSecurityManager = new DefaultSandboxSecurityManager(new DefaultSandboxPermissions());
     }
 
     @Before
